1. Moved post to separate component
